requirements.txt
python-dotenv==1.0.0
Faker==20.1.0
mariadb==1.1.8
